Got acceleration/deceleration to work
Look at MovePlayer function in the SonicMovement script.
First I tried Lerping, didn't work. I tried to do it by elapsedTime, but it was giving me weird explosive accelerations. Then I tried movetowards but the control was terrible. Initially I added drag, but that messed with the acceleration/deceleration. Then I stumbled upon this video and it's part 2:
https://www.youtube.com/watch?v=raIX3znrik4&t=1s

It explained that the rotateTowards makes the controlling better. ProjectOnPlane gives velocity the player is going in, in relation to it's ground, giving us horizontal velocity  Project was used similarly for verticalVelocity. technically, I can just use rb.velocity, but I plan on implementing slope movement, so we'll need this. I might refernce these 2 videos a bit more, but I'll restrain myself from doing it too much given that I'm trying to create it from scratch. It's good to use resources and references, but when showcasing ability, it's important to highlight, my own creative problem solving