Fixed it! Slope movement works. Time to polish it as well as apply rotation to the GFX
Four vertical velocity, which was our Vector3.Project, I only add it to our velocity if the player is not grounded, which would be to simply subtract our y velocity by gravity. Because this new method adds gravity, and does so in a much simpler way, I no longer have to use a constant force component to simulate gravity when the player wasn't grounded.
Because vertical velocity was only applied when in the air, and our horizontal velocity already moves our velocity towards our desired velocity, we can still move on upwards
However, this solution caused a problem. When the player would jump, sometimes, the player abruplty stops rising right after pressing the jump. This indicated that when the player jumped, the player tried to stick itself to the ground and would apply gravity. To fix this, we simply make sure that not only would we need to check if the character is grounded, but also if they are ready to jump (this check goes with sticking the player to the ground AND when considering vertical velocity). readyToJump is a bool that is only true for a fraction of a second right after the player jumps. This is done so that the ground checking ray doesn't accidently mark the player as grounded right after jumping. It gives the player enough time to reach a big enough height so that the ray can't accidently mark the player as grounded for the first few frames after leaving the ground. By adding this check, we make sure the player reaches enough height were the system won't accidently mark the player as grounded, and therefore, apply gravity/stick player to the ground.
Now that that is solved, time to polish the slope movement a bit more. I need the player to rotate towards the normal of the surface as well as shoot the grounding ray  towards the surface the player is on, rather than just straight down.