Finally got it to work. I did one last technique which was to manually apply the speed to the horizontalVelocity so that the player doesn't lose speed
By storing the previous speed (horizontal.magnitude), we can compare it to out current speed, and make the final speed to whichever of the 2 is higher. We can do so by Normalizing the vector, which gives it a value of 1 in the direction we want to go in, then just multiply it by our desired speed, which would give it the magnitude we want.

Now, all that's left for the ground movement is to make the player acc/decc on slopes
After that, on to the air movement (which shouldn't take too long)
At that point, I'll have the general movement down. Which means I'll start doing the custom movemet like rail grinding, boost, spindash, etc