Ground Movement works now with velocity going down to a certain point when going up hill and velocity going up to a certain point going downhill. What I did was simply check if the player was going downhill/uphill/flat surface. Once we have that, simply change the desire speed based on said surface movement. And for it to work with the current system, I simply made it so that if the speed is higher than what it's supposed to be, than simply keep using said speed until it inevitably reaches the speed it needs to
I also started working on the air movement.
Before I start working on that though, I need to fix a bug.
When the player is above the ground, it snaps the player towards the ground even when we're still a decent amount above it.
It now works after we jumped, but when the player simply runs off a slope, the early sticking will still happen.
Once I solve that. It shouldn't be too hard to implement the things I plan for the jumping (air acceleration and decceleratoin, short hops, max air speed)