Almost done fixing the early ground sticking bug
For the most part it's fixed.
I debugged to see what was going on, and I noticed that whenever the unsticking would happen, it would be because on the moment of transition from surface to different angled surface, the distance from the player to the ground would grow larger than the threshold.
So the rays and the checks work.
So what's going on is that the player travels really far during frame updates when transitioning from surface to surface, that by the time the new ground check happens, the player already traveled too far.
So to fix that, I made sure to move everything ground related to fixedupdate because it's more consistent and plays with the physics engine, which is where our player is supposed to thrive on.
Now, it works much better. It still a bit inconsistent at times though, so to fix that, instead of using a raycast, I think I'm going to use a boxcast or spherecast to give the grounding a bit more of a forgiving ground detection without having to raise the ground threshold (which if I do, the player will stick to the ground too early).
PROGRAMMING:
The method used for solving this was simply shoot a long ray. If it doesn't hit, make the player ungrounded which will make the player not stick.
If it does, measure the distance between the hitpoint and the player. If it's a small amount, the player is supposed to be grounded. Otherwise, the player is ungrounded.
Like I said, still needs small tweaks, but for the most part. It works.