I realized that the reason why those small things were happening was because of the way I put some platforms. Looks like it works now!
The player stick will work with any surface from any surface as long as the change in angle isn't to big or there is rounded edge-like transition, which is good because we don't wan't the player to be able to launch themselves from sharp angles. That makes sense and feels good. A rounded edge will for the most part mean, "I can traverse through this as long as I keep up speed" and a sharp edge of a platform or a surface should mean "If I run off, I'll fall".
To explain it further, Imagen the player is on a platform. If the player runs towards the edge assuming it's sharp, you expect the player to run off, not stick on the side of the platform.
In addition, this won't interfere with the jump.
Now, onto the air acc/dec. Once we have that done. Homing attack.
I'll leave boost spindash and unsticing from a wall for later seeing as how those are more so special parts of the moveset rather than things the basics