Short hopping done. Now, fix last jump bug
The way it works is simply when the player presses the jump key, run a small timer. If the player released the button within the timer, they want to short hop. Otherwise, they want to full hop.
After we know what the player wants, jump.
I restructured the jump aspect of the code. More specifically. I moved everything that was for jumping into a coroutine. This way, not only do we make sure the jump doesn't get called until it's determined whether the player wants to short hop or not. Not only that, we can take advantage of the waitforseconds return call from a coroutine to reset the readyForJump, which is good because coroutines are better for performance when compared to invoking functions
To be honest, I like this structure more. Now MyInput function really does look like it's just input detection and all jump logic is neatly organized outside of it.
Now, while the jump mechanics are for the most part done, there is one last bug. When jumping in slopes, if the player is going downhill the player will zoom far forward rather than just jumping. Also, when jumping while going uphill, the player will jump too high. I like the player jumping higher when going up hill because momentum is preserved, but not when it's too much to the point were it's exxagerated.
Once this is fixed, I'd say homing attack is next.