Finished jumping in slope movement
Before, when we pressed the jump button, for a small moment, gravity would be applied right before the jump happened, so this made the player only go up a tiny bit. Also, because readToJump became false because of jumping, this would make the player ignore the groundsitcking and mess with the horizontal velocity. As for jumping up slopes it was too strong because again, for a few frames, when jumping, the vertical/horizontal velocity would be affected.
The way it got fixed is as follows:
Once the player wants to jump, for a fraction of a second, ignore the regular ground detection and reset readytojump and grounded, that way, by the time the actual jump happens, the jump isn't affected by the vertical velocity checks in moveplayer.
Additionally, during this timer check if the surface under the player is different than the one the player started the jump in. This is done because if the player just waits for the readyToJump function to reset the jump, then when jumping from a surface to a differently angled surface, if the jump ended before we're readyToJump, the groundsticking will be ignored, and the player will just bounce off like a ball because of the collider. If the surface below is different, or if there is no surface, then we can exit this timer because we have exited the few frames window that allowed impact for the vertical velocity
Lastly, to make the jump feel better, and keep the feel of momentum, I made the jump always be in the direction of the surface the player is standing on (with a bit of a higher angle for when going up a slope). This helps the feel that the character is actually able to interact with the environment.