Quick recap of what we have (ground movement, including slopes, jumping, short hopping, slope movement and jumping has physics simulated into it). Might polish some variable numbers
So what we have right now is: Ground movement (going up/down slopes makes you go slower/faster, Air movement (jumping, jumping in slopes works now and the force and direction is relative to the surface. When jumping up hill, the player will leap up realitive to the speed and when jumping down hill the player will jump in direction of the normal of the surface, leading to cool momentum tricks) and short hopping (which keeps speed as a reward, Normal jumping slows the character down) and ground sticking works as intended.
I might test out the game a bit more to tweak a few numbers to make the game feel as best as possible.
I'm deciding whether the next step is either homing attack or boost