To fix the problem mentioned in the last commit, I decided to just use a trigger collider. It's simpler and uses less messy looking code
Instead of making calculation, I just simply made a trigger collider around the player's length (that barely doesn't cover the feet) and turned it on whenever we were in this jump interval for keeping ground in false, and turned it off for whenever the time interval was no longer running, or if the collider touch the ground. If the trugger touched the ground, turn off the timer and the trigger collider. It works nicely. I thought about using one before, but I didn't want to because I wanted to see if it was simple through code. At one point, I tried a few too many things that didn't properly work and thought that to not waste time, it's probably best to just use a trigger collider
I did think about using the dot product between the velocity of the player and the surface detected during the timer, and if the result is negative, or even close to zero, then restick. This may have worked with some tweaking, but the solution I have right now already works and is simpler.
If by any chance, the collision is odd during this time interval, tweak the collider. It should be fine, but it's worth noting
Now, we can continue to work on the homing attack