Homing Attack implemented! On to spindash
The way it goes is as follows. The player can only throw a homing attack if they are allowed to, and if they are in the air. Once the requirements are meant and the player clicks the homing attack button, we calculate the target for the homing attack based on how close they are to the center of the screen AND how close they are to the player (and if they are within a certain radius of the player). Then transition into the homingattack state
Then, if a target was determined, make the character travel quickly towards the target. Once they are extremely close, apply force upwards to make it look like the player is bouncing off the enemy, and transition into the regular movement state (functionality and variables have been made for future use of how much momentum the character keeps after the homing attack based on the previous direction inputted).
If an target was not found, simply shoot the player based on the direction the player last pressed (the way the player would be facing).
I'm glad that this, for the most part, was simple and quick compared to what I've previously done.
Onto the spindash!!!