Started the Spindash logic
If the player inputs the spindash key, freeze the controls and charge spindash. I chose lerp instead of movetowards because I want the player to gain a lot of charge in the beginning, and little in the end. This makes using the spindash quick nice, but not as satisfying as holding it. Having said that, because it gives a lot of charge in the beginning, it will still be good