Finished Spindash! On to Boost
The way it works is like Sonic Adventure 1 and 2.
The logic as to how to do it is really similar to regular movement, just with some variable and logic tweaks. So to understand it, look at the comments and commits that talk about the regular movement (MovePlayer).
Having said all this, the code looks a bit messy, so I'll clean it up before starting the boost.
Once I finish the boost, all the sonic movement mechanics will have been done. Which means that at that point, on to level design and animations.