small script clean up. On to Boost!
Just polished the script a bit in regard to the spin-dash. I might clean it up a bit more later or separate it into different scripts some other time
I think I might add a mechanic were you can charge the spin-dash in the air while moving. I'll think on it a bit more and if it the idea still sounds cool by the time I finish, I'll implement it