Finished Boost! I'll tidy up the details later
Boost is done, but I wanna adjust details a little later. I also added the simple ui asset from the asset store and made a boost meter, that works, for the boost. To do it, I followed this video:
https://www.youtube.com/watch?v=AyuQXfgVk3U

I also made a small ring that will add to the boost later and rotate

I'll explain the boost in the next commit