Boost working good and added rings. On to polishment and level designing
Fixed a small transition issue there was with the spindash, as well as added "rings" that refill the boost meter. Here are the credits for the model:
"Sonic Ring" (https://skfb.ly/6DpE9) by Shuaib_Ahmed is licensed under Creative Commons Attribution (http://creativecommons.org/licenses/by/4.0/).
When I finish this project, I'll make a text document listing all the credits.
I know I said I was going to explain the boost, and I will in the next. Sorry for the delay!