Explaining Boost and added model
The boost is pretty simple. Simply makes the players velocity magnitude equal to boost speed during this state. As for the meter, lower it as it's being used. The transitions, thanks to using state machine, were pretty easy to handle as well.
As for the model and animation. I found a model online WITH animations!
Here's the credit for the model:
"Sonic Forces Sonic Rig" (https://skfb.ly/oZEV8) by NO DONT EAT ME CASEOH (Ferris wheel) is licensed under Creative Commons Attribution (http://creativecommons.org/licenses/by/4.0/).
To make it clear, I did not make this model or animations nor do I plan on selling this showcase I'm making. Anything I do in this project is for showcasing purposes to showcase my ability in unity and all things I use in this project to help portray that, that I did not make, will be given their credit to the great people that made them.
If there is any issue with me using this model, please let me know to whoever is reading this. I'll also put this type of note in the read me.
For the model, right now, I simply have a script that appropriately updates the rotation and position of the model based on the player objects position and rotation. I'll start implementing the animations next time I get on.