Setup the basic animations
All the regular movement animations, jump, and falling animations have been implemented.
They were all straightforward. The only one that was a different was the ground movement animations. For idle, walk, run, and dash, I made a blend tree. The way it works is that the animation that will play will be dependent on the speed. If the speed is between two animations thresholds, then the animation that will play will be a blended animation between the 2 where how much of each animation is played is determined by how close the determining value is to the threshold values for the animations.
This video helps understand the concept as well as explains how to set it up:
https://youtu.be/DnqcxhOV-xQ?si=YPB85Ma1Igq8GTxu