Still working on rail grinding
We can get on the dolly cart now. The extra check in stomp is so that when the player stomps down from right above, the player collider and the rail collider don't accidently skip each other for going to fast. Next session, I'll work on moving the player once the player is on the dolly cart