Fixed a bug on getting on the rail. Getting on and off rail by jumping works good
When stomping, the player would move so fast that sometimes the collision between the player and a rail would be ignored, even when the player's collision mode was in continuous/continuous dynamic (the next frame position of the player would be past the colliders).
Initially, I thought I could do something similar to that of the grounding (make a large ray downwards and if the ray hits a rail within a short distance, stick to the rail). While this did make it better, the problem would still happen here and there.
So I decided to try to simulate the actual next position, but on this current frame in order to get more accurate results (using linecast because of start and end position). This allows us to check the next frame without having to wait for the next frame. It worked. While I understood the problem and the required solution, the actual implementation was made easier thanks to this video which actually was going over a similar problem. Here's the video link:
https://www.youtube.com/watch?v=PD5RzsAEHmw