Rail Grinding works for the most part, but still needs a lot of polish
The it works is like so. We have one empty object as a parent of the Rail's body/collider, the spline for the rail path, and the cart for the spline. We do this to keep individuality between rails.
As I said before, I used Cinemachine spline and dolly cart tools. The body/collider object has a script that when the player touches the object, the script will move the dolly cart towards where the player landed in (it finds the spot by checking the distance from the player at each point of the spline through binary search), and initializes the cart on the SonicMovement script. Also, it determines the start speed the rail as well as the direction the player should move in along the rail (towards or away from the endpoint). Lastly, it switches the players movementState to Railgrinding.

As for the Sonic script, the Railgrinding function simply moves the cart, and then moves the player to the location of the cart, simulating rail movement.
The checks for leaving and entering the rail are due to the rail collider sticking the player back on the rail when the player wants to get off. To solve it, we simply ignore the rail for a brief moment.
I'll make sure to simplify the code once the rail movement is completely done, because as I said, the rail movement isn't finished.

What the rail movements is missing is the following:
- Exiting rail once an end is reached
- When the player exits the rail, whether by jumping or leaving, conserve the speed, keeping the momentum
- Make the rotation of the player match the rails normal
- Animation for rail movement (right now, it's a T-pose lol)
- Boost capabilities in rail
- Moving from rail to rail
- Test out with complex-shaped rails to make sure it works. It should work, but just to make sure

What we have:
- Entering/Exiting rail through jump
- Momentum on rails based on the speed the player enter the rail at as well as direction of rail movement.
- rail system

Still got quite a bit of things left, but the overall system and foundation is done