Finished Rail exiting with momentum
Finished:
- Exiting rail once an end is reached
- When the player exits the rail, whether by jumping or leaving, conserve the speed, keeping the momentum

Still need:
- Make the rotation of the player match the rails normal
- Animation for rail movement (right now, it's a T-pose lol)
- Boost capabilities in rail
- Moving from rail to rail
- Test out with complex-shaped rails to make sure it works. It should work, but just to make sure

The way it works is simply, when exiting the rail, make the linear velocity (rb.linearVelocity/rb.velocity for earlier versions) equal to the tangent of the point sonic is on the rail, keeping the momentum. Also, if the player reached the end of the rail (for splines it was if the next position is less than 0 and more than the end length), and disconnect.
The main stuff is honestly done at this point, but adding animation and rotation to this is what's going to make it look really pretty, as well as effects of course