Rotation on rail works. Also, movement for rails is now physics based. On to making an animation for it
I used Full Frenet Frames to solve for the up rotation. That plus the forward given to us by the tangent results in the correct orientation. Might use this for the homing attack for smoother transition.
Also, movement for the rails is now physics based becuase moving with transform at high speeds results in jittering. I first put it there for testing purposes, but now, with the physics movement, the movement is now smooth