Character now faces towards the right of the way they are riding the rail
I used a matrix rotation setting method, which basically allows you to set the forward, right and up manually for a rotation. Because I already had the Up and forward, I used there cross product to get the right. After that, I simply made it so that when the player is on a rail, the models orientation is dictated by all rotations. Lastly, after a bit of testing and tweaking the values for the matrix, we got the rotation going correctly.
Now, rail functionality is officially completed. Only polishment is needed. In other words, we only need:
- Animation for rail grinding
- Mesh for the rails (right now, even though they work, they aren't visible unless gizmos is on)