Animation for rail made and implemented
I'm not an animator but I need the rail animation, so I decided to whip up a rail grind animation pose. It looks a lot better now. The pose isn't completely perfect, but it gets the job done really well, not it feels like he's actually grinding on the rail.
I may later ask a friend that now's blender if he can help make a few adjustments to the rail grind pose/animation, but that will be all the way by the end, and that's if this rail pose doesn't feel right, which I don't think will happen because like I said, it get's the job done and it feels good.

Now, I'm going to see if I can make a mesh or find a rail model so that the rails are seen without gizmos