Polished jump while touching a wall and boost meter clamp
Before when touching a wall, if you jumped, then the player would technically detect ground, because of the wall, and stick the player to the ground, so the player wouldn't jump. Now, the player jumps nicely while touching the wall, I basically made a collider that when the player touched the wall when the jump started, then don't stick.
As for the boost meter, I made it so that the boost meter can't fill above 100%