Player is no longer able to pass through objects.
Because when we sticked the player to the ground it was done with transform, if an obstacle was skinny enough, than the player would be able to pass through it because of it's high velocity pushing plus setting the position.
To fix it, we simply shoot a ray in the direction the player is running in. If it hits something, then that means the player needs to crash, so we set the rigidbody velocity, and the horizontal velocity variable, to 0.
Right now, the code isn't commented for it. I'll comment it tomorrow because right now I'm tired.