I decided to make 3 rocks instead of 4
The purpose of the rocks are so that the player experiences the different movement types, with exceptions to things like stomp and slide which are more tools to traverse a level rather than momentum experiences. With the first, it lets the player know what the rocks look like. The second shows momentum, rails, and homing attack. The last one should show boost and spindash. These 2 should be shown not really as necessities, but more so fun options. So I think I'll make a huge hill, that the player run up, via boost, and go down, via spindash