I realized the 3rd was just there right in front of the player. So I decided to take it off. FOREST ZONE FINISHED
Finishe dthe forest section. It mainly seerves as a place where the player can play around with the momentum and the rails, which I believe does nicely. As a reiminder to anyone viewing this. The purpose of this is mainly for gameplay, so while I do like how the forest looks, I think it's best I move on to the level considering that spensing more time on the forest means more overall world/level design, which is nice, but I mainly want to focus on the gameplay.

What's next?
- Level
- More level object (dash panel, ramp, etc)
- Speed check for going at a very vertical angle
- Loop-de-loop polish
- Camerawork (optional)
- VFX (optional)

Once this top things get done. I believe that will be it for this project. We're almost there!
By the way, I'll make sure to leave instructions in the readme as well as credits