Made functionality of speed checking the player at high angles. Also light spline realization lol
I did not know that about spline's instantiate component so there's a chance I may not have needed to do generative meshes and colliders for the rails lol. But I'll keep the rails how they are for 3 reasons
1) they get the job done
2) If anyone wants to use spline instantiation, they'll be able to follow similarly to the steps I took with the loop de loop because I'm using that for the loop de loop.
3) Time

I may decide to change that in the future, but for now, they'll stay the way they are.
learned lesson:
search all the resources available for an asset before using it!