LoopDeLoop's are now functional and are prefabs
As I mentioned before, I used a spline, and the instantiate component, to make the loop de loop itself
The way the functionality works is as follows:
When the player is on the loop, the spline cart follows the player as closely as possible.
Then, based on the camera angle, we get the tangent of the cart's position on the spline. This gives us the forward direction for the loop, or in other words, the path we want to take.
After that, we just simply make the orientation's forward for the player match that tangent.
Note that we do not invert the tangent's direction after a certain angle because the tangent does a full loop. Meaning, that once the player has rotated a certain amount throughout the loop, the loop will face differently based on the camera than when the player started. So if we don't stop inverting the tangent at those points, the tangent will be different when going from end to beginning once we hit that section, resulting in the orientation abruptly changing.
I added a toggle that makes it easier on developers for some occasions which is called OnlyForwardDirection. If turned on, the tangent doesn't depend on camera angle and the tangent is always put as if the player is always traversing from beginning to end of spline knots. This makes it simple to make cool camera movements throughout the loop. Without it, the loop should still work, the camera for the most part is in it's regular state.
If anyone wants to modify the tangent orientation, they can simply make it so that tangent equals orientation.directionyouwant. For example, for 2D sections, orientation.right = tangent might be the way to go. I won't do it here because I want to focus on a more sonic adventure sort of style, but I wanted to jot that down in case anyone wants to modify. I may make it an enum later, but that will require much more testing, as well as assuming I did a camera work system that brings in 2D aspects (I don't think I will because again, that's not what I'm trying to go for).
This is how the loop works. Now that speed checking and loops are done, On to Level Designing! (level design tools, then level).