Bumpers and Dash panels finished
In terms of logic, neither was difficult. For bumper, simply impulse the player towards a direction and once they slow down a certain, adjustable, amount, the player can move. Similarly for the dash panel, except that instead of impulse, it's keeping the speed at a certain point for a brief, adjustable, moment.
What was a bit on the harder said was everything that surrounded that (animation, player model sticking close to colliders, managing which bumper/panels info to use). To know which bumper/panels' info to use, I simply keep track of the current bumper/panel in use.
I also needed to do a bit of setup for each before the actual effect would start appropriately.
I explained everything here in terms of logic and all niche setup stuff is in comments all also add a few more comments to it later, so I think I'll skip that stuff here because it's more so details that surrounded the method rather than part of the formula. So if anyone wants to make a variant of one of these, the main idea is the one they need to follow for general purposes, which I explained above.
This really ties back to what it means to be a gameplay programmer. Not just worrying about functionality, but how it affects the other aspects of the gameplay such as animation, other gameplay loops, environment, etc., all while maintaining workflow.
So to sum up, the functionality of both was rather simple, but the setup, clean up, animation, and workflow maintenance was the challenge, which I'll humbly say I think I did good :)
Now that this is over, time to work on the level! Almost done!
Once I finish the level I'll see if I'll work on the VFX and camera work. I'll give my thoughts on those some other time.