Added camera work to bumpers
Similar to dash panels, the camera has the option to face the direction of the bumper. Also, the bumper has the additional functionality to face downwards (helps when sending the player upwards).
When the player hits the bumper/dashpanel, pass along the desired direction towards the new camera method. That will move the current camera angle towards the desired one through spherical interpolation (moves towards target in a circular fashion).
Once that ends, it'll update the new rotX/Y to fit this angle.