Fixed transition for going into/out a loop
I used a new spline as the path the camera would take. The spline postion of the camera would be the same as the spline position of the player on the loopdeloop. After that, simply transition the camera back to the player after exiting a loop de loop. I made the transitions of going in and out of a loop coroutines so that it would smoothly transition in and out of the loop properly. Keeping track of the position, rotations, and cleaning up the camera follow script so that it's easier came in very handy.

I also made slight adjustments to the level to make it shorter. Just need to do the camera collision and that should be it for the camera work!