Camera Collision script decision
Alright so in the game, from the beginning, I had a camera collision script. It worked, and decided to leave it off to the end because I didn't know just how much camera work would be done. Now that the camerawork is done, I though about going back in and checking it out. The script works BUT it works differently in the forest than in the level. This is because of the different style. And it actually works alongside the camerawork stuff of the level, so there is no conflicting issues there. This makes me realize that between the adventure, boost, open zone, or free roam (common in fan games) style, the camera collision is honestly pretty different. So I think the best decision is to keep this base camera script as it is and allow others adjust it the way they want it. Here are some tips for each that I would apply:

Open zone (forest-like) and free roam
- Keep it the same, but make sure it doesn't collide with anything you don't want it to collide with like a rail
Adventure
- Have a global state of when/where the camera collision is allowed. This is in order to not conflict with any cinematic effect done
Boost
- It depends how deep you want to go with the boost formula. If you want to go super deep like as close as possible to something like Unleashed, than give the main camera a more stretched/narrow effect when running at high speeds, and make sure that the camera doesn't exceed any structure that surrounds sonic (boost games have a lot of vertical structures to the sides of sonic). If you want to go a bit less deep, like in frontiers, I'd suggest following something similar to the adventure tip.

The base is pretty flexible due to it's simplicity, and while is nice to have, it's not completely necessary depending on how you build your levels and control your camera. Even so, feel free to modify anything to your needs :)

With that being said. The next part is to make sure all the things in the level, which are the final versions of the project, work in the forest and sandbox. They should work, but it's always best to make sure. I'll also erase any unused prefabs/scripts. I want to do that before the VFX